Homepage table working correctly

+++ feedtrac/classes/LoginContr.class.php
@@ -107,6 +107,11 @@ class LoginContr extends Login {
  // Get the user ID from email
    public function get_userid_email($email){
 
+    // Check if any input is empty
+    if ($this->empty_input_check($email)){
+        header("location: login.php?error=BadSQLQuery");
+        exit();
+    }
   
     return $this->get_id_email($email);
 }
@@ -114,8 +119,12 @@ class LoginContr extends Login {
 // Get the token using userID
    public function get_token($userID){
 
+    // Check if any input is empty
+    if ($this->empty_input_check($userID)){
+        header("location: login.php?error=BadSQLQuery");
+        exit();
+    }
   
-
     return $this->get_recovery_code($userID);
 }
 
@@ -141,12 +150,23 @@ class LoginContr extends Login {
 
     public function delete_recovery_record($userID){
        
+        // Check if any input is empty
+        if ($this->empty_input_check($userID)){
+            header("location: login.php?error=BadSQLQuery");
+            exit();
+        }    
+
         return $this->delete_record($userID);
-        
     }
 
 
     public function update_user_password($hashed_password,$userID){
+
+        // Check if any input is empty
+        if ($this->empty_input_check($hashed_password,$userID)){
+            header("location: login.php?error=BadSQLQuery");
+            exit();
+        }    
  
         return $this->update_password($hashed_password,$userID);
     }
@@ -171,7 +191,7 @@ class LoginContr extends Login {
 
         // Send email
         /* mail($email, $subject, $message, $headers);*/
-        echo $message; // DELETE THIS
+        echo $message; // THIS IS ONLY UNTIL WE HAVE A WORKING EMAIL SERVER (just a proof of concept as we can't send emails from localhost)
 
         return true;
         
@@ -180,56 +200,30 @@ class LoginContr extends Login {
     // Get username
     public function get_username(){
 
-        //prepare the SQL query
-        $stmt = $this->connect()->prepare("SELECT username FROM user WHERE userID = ?");
-
-        // Check if the SQL query is valid
-        if(!$stmt->execute([$_SESSION['userID']])){
-            header("location: settings.php?error=BadSQLQuery");
-            exit();
-        }
-
-        $result = $stmt->get_result()->fetch_assoc()['username'];
-        return $result;
+        return $this->get_username_id($_SESSION['userID']);
     }
 
     // Delete Account
     public function delete_account(){
 
-        //prepare the SQL query
-        $stmt = $this->connect()->prepare(
-            "UPDATE user
-            SET email = 'deleted@example.com',
-                username = 'deleted_user',
-                passwordHash = 'deleted_password',
-                fName = 'deleted_fName',
-                lName = 'deleted_lName',
-                pronouns = NULL,
-                position = 'deleted_position',
-                activeAccount = 0
-            WHERE userID = ?"
-        );
-
-        // Check if the SQL query is valid
-        if(!$stmt->execute([$_SESSION['userID']])){
-            header("location: settings.php?error=BadSQLQuery");
-            exit();
-        }
+        return $this->get_username_id($_SESSION['userID']);
     }
 
-    //Check to see if account has been deleted
-    public function is_active($username){
 
-        //prepare the SQL query
-        $stmt = $this->connect()->prepare("SELECT activeAccount FROM user WHERE username = ?");
 
-        // Check if the SQL query is valid
-        if(!$stmt->execute([$username])){
-            header("location: settings.php?error=BadSQLQuery");
-            exit();
-        }
+    // //Check to see if account has been deleted 
+    // public function is_active($username){
 
-        $result = $stmt->get_result()->fetch_assoc()['activeAccount'];
-        return $result;
-    }
+    //     //prepare the SQL query
+    //     $stmt = $this->connect()->prepare("SELECT activeAccount FROM user WHERE username = ?");
+
+    //     // Check if the SQL query is valid
+    //     if(!$stmt->execute([$username])){
+    //         header("location: settings.php?error=BadSQLQuery");
+    //         exit();
+    //     }
+
+    //     $result = $stmt->get_result()->fetch_assoc()['activeAccount'];
+    //     return $result;
+    // }
 }
